+++ voxygen/src/hud/util.rs
@@ -144,7 +144,10 @@ pub fn stats_count(item: &dyn ItemDesc, msm: &MaterialStatManifest) -> usize {
             Effects::Any(_) | Effects::One(_) => 1,
             Effects::All(effects) => effects.len(),
         },
-        ItemKind::RecipeGroup { recipes } => recipes.len(),
+        ItemKind::RecipeGroup { recipes } => {
+            /* Add one for recipe known/not known message */
+            recipes.len() + 1
+        },
         ItemKind::ModularComponent { .. } => 6,
         _ => 0,
     };